put PILCO testing code in a tmp branch

+++ barl/acq/acquisition.py
@@ -886,7 +886,7 @@ class KGRLAcqFunction(AcqFunction):
         model.set_data(data)
         model.initialize_function_sample_list(num_fs)
         current_states = np.repeat(current_states[np.newaxis, :, :], num_fs, axis=0)
-        current_states = tf.convert_to_tensor(current_states, dtype=tf.float32)
+        current_states = tf.convert_to_tensor(current_states, dtype=tf.float64)
         f_batch_list = model.call_function_sample_list
         returns = 0
         for t in range(rollout_horizon):
