+++ barl/envs/pilco_cartpole.py
@@ -218,13 +218,13 @@ def tf_get_pole_pos(x):
     theta = x[..., 2]
     pole_x = CartPoleSwingUpEnv.POLE_LENGTH*tf.sin(theta)
     pole_y = CartPoleSwingUpEnv.POLE_LENGTH*tf.cos(theta)
-    position = tf.cast(tf.stack([xpos + pole_x, pole_y], axis=-1), tf.float32)
+    position = tf.cast(tf.stack([xpos + pole_x, pole_y], axis=-1), tf.float64)
     return position
 
 
 def tf_pilco_cartpole_reward(x, next_obs):
     position = tf_get_pole_pos(next_obs)
-    goal = tf.constant([0.0, CartPoleSwingUpEnv.POLE_LENGTH])
+    goal = tf.constant([0.0, CartPoleSwingUpEnv.POLE_LENGTH], dtype=tf.float64)
     squared_distance = tf.reduce_sum((position - goal)**2, axis=-1)
     squared_sigma = 0.25**2
     costs = 1 - tf.exp(-0.5*squared_distance/squared_sigma)
